refactor : 안 쓰는 파일 지우기
