+++ src/main/java/roomescape/member/right/AdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.login;
+package roomescape.member.right;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -6,6 +6,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
+import roomescape.login.JwtUtil;
 import roomescape.member.Member;
 import roomescape.member.MemberDao;
