+++ src/main/java/roomescape/member/right/WebConfig.java
@@ -1,4 +1,4 @@
-package roomescape.login;
+package roomescape.member.right;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
